--- Procfile
@@ -1,1 +1,0 @@
-web: bin/go-jwt-auth
