+++ README.md
@@ -1,0 +1,9 @@
+# Go Auth Microservice
+
+## Inspired by the work of:
+
+## d-vignesh's go-jwt-auth -> [d-vignesh/go-jwt-auth](https://github.com/d-vignesh/go-jwt-auth)
+
+## and
+
+## obe711's node-server-boilerplate -> [obe711/node-server-boilerplate](https://github.com/obe711/node-server-boilerplate)
