--- data/models.go
@@ -1,34 +1,0 @@
-package data
-
-import (
-	"time"
-)
-
-// User is the data type for user object
-type User struct {
-	ID         string    `json:"id" sql:"id"`
-	Email      string    `json:"email" validate:"required" sql:"email"`
-	Password   string    `json:"password" validate:"required" sql:"password"`
-	Username   string    `json:"username" sql:"username"`
-	TokenHash  string    `json:"tokenhash" sql:"tokenhash"`
-	IsVerified bool      `json:"isverified" sql:"isverified"`
-	CreatedAt  time.Time `json:"createdat" sql:"createdat"`
-	UpdatedAt  time.Time `json:"updatedat" sql:"updatedat"`
-}
-
-type VerificationDataType int
-
-const (
-	MailConfirmation VerificationDataType = iota + 1
-	PassReset
-)
-
-// VerificationData represents the type for the data stored for verification.
-type VerificationData struct {
-	Email     string    `json:"email" validate:"required" sql:"email"`
-	Code      string    `json:"code" validate:"required" sql:"code"`
-	ExpiresAt time.Time `json:"expiresat" sql:"expiresat"`
-	Type      VerificationDataType    `json:"type" sql:"type"`
-}
-
-
