+++ main.go
@@ -2,141 +2,63 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 	"os"
 	"os/signal"
+	"syscall"
 	"time"
 
-	"github.com/d-vignesh/go-jwt-auth/data"
-	"github.com/d-vignesh/go-jwt-auth/handlers"
-	"github.com/d-vignesh/go-jwt-auth/service"
-	"github.com/d-vignesh/go-jwt-auth/utils"
-	"github.com/gorilla/mux"
-	"github.com/hashicorp/go-hclog"
+	"github.com/caseyrwebb/go-jwt-auth/app"
+	"github.com/caseyrwebb/go-jwt-auth/app/utils"
+	"go.uber.org/zap"
 )
-
-// schema for user table
-const userSchema = `
-		create table if not exists users (
-			id 		   Varchar(36) not null,
-			email 	   Varchar(100) not null unique,
-			username   Varchar(225),
-			password   Varchar(225) not null,
-			tokenhash  Varchar(15) not null,
-			isverified Boolean default false,
-			createdat  Timestamp not null,
-			updatedat  Timestamp not null,
-			Primary Key (id)
-		);
-`
-
-const verificationSchema = `
-		create table if not exists verifications (
-			email 		Varchar(100) not null,
-			code  		Varchar(10) not null,
-			expiresat 	Timestamp not null,
-			type        Varchar(10) not null,
-			Primary Key (email),
-			Constraint fk_user_email Foreign Key(email) References users(email)
-				On Delete Cascade On Update Cascade
-		)
-`
 
 func main() {
 
-	logger := utils.NewLogger()
+	filename := "logs.log"
+	logger := utils.CustomLogger(filename)
 
 	configs := utils.NewConfigurations(logger)
 
-	// validator contains all the methods that are need to validate the user json in request
-	validator := data.NewValidation()
-
-	// create a new connection to the postgres db store
-	db, err := data.NewConnection(configs, logger)
+	app := app.New(*logger, configs)
+	app.DB.SetDBLogger(logger)
+	err := app.DB.Open(configs)
 	if err != nil {
-		logger.Error("unable to connect to db", "error", err)
-		panic(err)
+		logger.Error(fmt.Sprintf("%s %s %s", "could not connect to the database", "error", err))
+		os.Exit(1)
 	}
-	defer db.Close()
 
-	// creation of user table.
-	db.MustExec(userSchema)
-	db.MustExec(verificationSchema)
+	defer app.DB.Close()
 
-	// repository contains all the methods that interact with DB to perform CURD operations for user.
-	repository := data.NewPostgresRepository(db, logger)
-
-	// authService contains all methods that help in authorizing a user request
-	authService := service.NewAuthService(logger, configs)
-
-	// mailService contains the utility methods to send an email
-	mailService := service.NewSGMailService(logger, configs)
-
-	// UserHandler encapsulates all the services related to user
-	uh := handlers.NewAuthHandler(logger, configs, validator, repository, authService, mailService)
-
-	// create a serve mux
-	sm := mux.NewRouter()
-
-	// register handlers
-	postR := sm.Methods(http.MethodPost).Subrouter()
-
-	mailR := sm.PathPrefix("/verify").Methods(http.MethodPost).Subrouter()
-	mailR.HandleFunc("/mail", uh.VerifyMail)
-	mailR.HandleFunc("/password-reset", uh.VerifyPasswordReset)
-	mailR.Use(uh.MiddlewareValidateVerificationData)
-
-	postR.HandleFunc("/signup", uh.Signup)
-	postR.HandleFunc("/login", uh.Login)
-	postR.Use(uh.MiddlewareValidateUser)
-
-	// used the PathPrefix as workaround for scenarios where all the
-	// get requests must use the ValidateAccessToken middleware except
-	// the /refresh-token request which has to use ValidateRefreshToken middleware
-	refToken := sm.PathPrefix("/refresh-token").Subrouter()
-	refToken.HandleFunc("", uh.RefreshToken)
-	refToken.Use(uh.MiddlewareValidateRefreshToken)
-
-	getR := sm.Methods(http.MethodGet).Subrouter()
-	getR.HandleFunc("/greet", uh.Greet)
-	getR.HandleFunc("/get-password-reset-code", uh.GeneratePassResetCode)
-	getR.Use(uh.MiddlewareValidateAccessToken)
-
-	putR := sm.Methods(http.MethodPut).Subrouter()
-	putR.HandleFunc("/update-username", uh.UpdateUsername)
-	putR.HandleFunc("/reset-password", uh.ResetPassword)
-	putR.Use(uh.MiddlewareValidateAccessToken)
-
-	// create a server
 	svr := http.Server{
 		Addr:         configs.ServerAddress,
-		Handler:      sm,
-		ErrorLog:     logger.StandardLogger(&hclog.StandardLoggerOptions{}),
+		Handler:      app.Router,
+		ErrorLog:     zap.NewStdLog(logger),
 		ReadTimeout:  5 * time.Second,
 		WriteTimeout: 10 * time.Second,
 		IdleTimeout:  120 * time.Second,
 	}
 
-	// start the server
 	go func() {
-		logger.Info("starting the server at port", configs.ServerAddress)
+		logger.Info(fmt.Sprintf("%s %s", "starting the server at port", configs.ServerAddress))
 
 		err := svr.ListenAndServe()
 		if err != nil {
-			logger.Error("could not start the server", "error", err)
+			logger.Error(fmt.Sprintf("%s %s %s", "could not start the server", "error", err))
 			os.Exit(1)
 		}
 	}()
 
-	// look for interrupts for graceful shutdown
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt)
-	signal.Notify(c, os.Kill)
+	signal.Notify(c, syscall.SIGTERM)
 
 	sig := <-c
-	logger.Info("shutting down the server", "received signal", sig)
+	logger.Info(fmt.Sprintf("%s %s %s", "shutting down the server", "received signal", sig))
 
-	//gracefully shutdown the server, waiting max 30 seconds for current operations to complete
-	ctx, _ := context.WithTimeout(context.Background(), 30*time.Second)
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
 	svr.Shutdown(ctx)
+
 }
