--- notes.txt
@@ -1,7 +1,0 @@
-Generate rsa private key : 
-    openssl genrsa -out access-private.pem 2048
-    openssl genrsa -out refresh-private.pem 2048
-
-Export rsa public key : 
-    openssl rsa -in access-private.pem -outform PEM -pubout -out access-public.pem
-    openssl rsa -in refresh-private.pem -outform PEM -pubout -out refresh-public.pem
