--- utils/util.go
@@ -1,20 +1,0 @@
-package utils
-
-import (
-	"math/rand"
-	"strings"
-)
-
-const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-
-// GenerateRandomString generate a string of random characters of given length
-func GenerateRandomString(n int) string {
-	sb := strings.Builder{}
-	sb.Grow(n)
-	for i := 0; i < n; i++ {
-		idx := rand.Int63() % int64(len(letterBytes))
-		sb.WriteByte(letterBytes[idx])
-	}
-	return sb.String()
-}
-
